requirements.txt
openai==0.27.2
disnake==2.9.0
Pillow==9.5.0
requests==2.31.0

